+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Header keys must be lowercase.
 - Pseudo header values cannot be empty.
 - Fix off by one error for maximum size of frames.
+- Return protocol error for header or data sent after a stream reset
 
 ## [0.10.0](https://github.com/CrowdHailer/Ace/tree/0.10.0) - 2017-07-21